+++ project/game_components/ui/pseudo_button.h
@@ -27,7 +27,7 @@ struct Pseudo_Button
 
 /* Predef */
 
-struct Pseudo_Button create_pseudo_button(const char *const text, const SDL_Color inner_color, const struct Vec2 screen_pos, const unsigned int text_size, const unsigned int border_thickness, int *const exit_code);
+struct Pseudo_Button create_pseudo_button(const char *const text, const SDL_Color inner_color, const struct Vec2 screen_pos, const unsigned short text_size, const unsigned short border_thickness, int *const exit_code);
 void                   free_pseudo_button(struct Pseudo_Button *const target);
 
 #endif /// PSEUDO_BUTTON_H
