+++ project/graphics/fps.h
@@ -8,10 +8,10 @@
 #include "../logic/logic_layer.h" /// `time_tick_ns`.
 
 #define SEC_IN_NS 1000000000.0f
-extern const unsigned int fps_cap_options[4];
-extern unsigned int  curr_fps_cap_i;
-extern unsigned int  prev_fps;
-extern unsigned int  curr_fps;
+extern const unsigned short fps_cap_options[4];
+extern unsigned short  curr_fps_cap_i;
+extern unsigned short  prev_fps;
+extern unsigned short  curr_fps;
 extern bool          show_fps;
 
 #define FPS_POS_2_DIGITS  vec2(165, 5)
@@ -25,7 +25,7 @@ extern struct Vec2 fps_counter_position;
 struct FPS_Manager
 {
     bool         fps_capped;
-    unsigned int fps_cap;
+    unsigned short fps_cap;
     time_tick_ns        delta_ns;
     time_tick_ns target_delta_ns;
 };
@@ -34,7 +34,7 @@ extern struct FPS_Manager FPS_manager; /// Singleton.
 
 /* Predef */
 
-void set_fps_cap(const unsigned int new_fps_cap);
+void set_fps_cap(const unsigned short new_fps_cap);
 void rem_fps_cap(void);
 void wait_until_target_delta(void);
